Prevent self-xss in trace messages
Tracing probably shouldn't be turned on in anyone's production environment, even so, an action containing a malicious payload such as `'>"<svg/onload=confirm('xss!')>.index'` will execute when the trace is rendered on a page. I didn't see signs of other properties being affected, but it's possible. Don't know how likely it is that this would be exploited in a real world scenario, but encoding doesn't hurt.